[feat] move wallet connect location
