+++ packages/mobile/src/screen/setting/screens/general/general.tsx
@@ -65,21 +65,6 @@ export const SettingGeneralScreen: FunctionComponent = observer(() => {
 
           <PageButton
             title={intl.formatMessage({
-              id: 'page.setting.general.manage-WC-title',
-            })}
-            endIcon={
-              <ArrowRightIcon
-                size={24}
-                color={style.get('color-text-low').color}
-              />
-            }
-            onClick={() =>
-              navigate.navigate('Setting.General.ManageWalletConnect')
-            }
-          />
-
-          <PageButton
-            title={intl.formatMessage({
               id: 'page.setting.general.manage-non-native-chains-title',
             })}
             paragraph={intl.formatMessage({
