+++ packages/mobile/src/screen/setting/screens/security/wallet-connect/index.tsx
